skriveleif

+++ repository/src/main/kotlin/no/nav/aap/behandlingsflyt/integrasjon/datadeling/ApiInternGateway.kt
@@ -85,7 +85,7 @@ class ApiInternGatewayImpl() : ApiInternGateway {
                         no.nav.aap.behandlingsflyt.kontrakt.datadeling.TilkjentDTO(
                             tilkjentFom = tilkjentPeriode.periode.fom,
                             tilkjentTom = tilkjentPeriode.periode.tom,
-                            dagsats = tilkjentPeriode.tilkjent.dagsats.verdi,
+                            dagsats = tilkjentPeriode.tilkjent.dagsats.verdi.toInt(),
                             gradering = tilkjentPeriode.tilkjent.gradering.prosentverdi(),
                             grunnlag = tilkjentPeriode.tilkjent.grunnlag.verdi,
                             grunnlagsfaktor = tilkjentPeriode.tilkjent.grunnlagsfaktor.verdi(),
